feat: converting variants to constants, refact dynamic routing
