+++ projects/demo/src/app/strategies/custom-title.strategy.ts
@@ -2,8 +2,8 @@ import { inject, Injectable } from '@angular/core';
 import { RouterStateSnapshot, TitleStrategy } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
 
-const APP_TITLE_SUFFIX = 'i-Cell Design System';
-
+export const APP_TITLE_SUFFIX = 'i-Cell Design System';
+const TRANSLATION_KEY_REGEX = /^[A-Z_]+(\.[A-Z_]+)*$/;
 @Injectable({ providedIn: 'root' })
 export class CustomTitleStrategy extends TitleStrategy {
   private readonly _translate = inject(TranslateService);
@@ -35,6 +35,6 @@ export class CustomTitleStrategy extends TitleStrategy {
   }
 
   private _isTranslationKey(str: string): boolean {
-    return /^[A-Z_]+(\.[A-Z_]+)*$/.test(str);
+    return TRANSLATION_KEY_REGEX.test(str);
   }
 }
